food-app/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
confluent-kafka==1.9.2
Django==4.1.2
fastavro==1.7.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jsonschema==4.16.0
kafka-python==2.0.2
pkgutil_resolve_name==1.3.10
protobuf==4.21.9
pycodestyle==2.9.1
pyrsistent==0.18.1
python-schema-registry-client==2.4.1
pytz==2022.5
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.10.0
